requirements.txt
# Core
python-dotenv
pydantic
typing_extensions

# API
fastapi
uvicorn

# LangChain & Graph
langchain
langchain-core
langchain-community
langchain-openai
langgraph
langchain_tavily
langchain-groq
tavily-python
langchain-qdrant
# Vector DB
faiss-cpu

# PDF & Images
pymupdf   # fitz
pillow

# OpenAI client
openai

# Utilities
requests


python-multipart


qdrant-client